requirements.txt
pandas
scikit-learn
ahocorasick-py
openai
kaggle
tqdm
python-dotenv
spacy